Audit fix wave 1 (P0): area registration, link split-brain, dead EW surface, rollback strand, publish doctrine, real stocks
From the full-system linkage audit (DMIS-LINKAGE-AUDIT.md, 154 findings,
ledger in DMIS-AUDIT-FIX-LOG.md — 7 items closed this wave, all live-verified):

- User Management: create/update accept + validate region/district/agency/
  stakeholder attachment (district must belong to region); modal gains
  role-driven pickers and an Area column; area officers can finally be
  registered from the product (verified: API-created RAS saw exactly its
  region's incidents; cross-region district rejected 400).
- Stakeholder link split-brain: users.stakeholder_id and stakeholders.user_id
  now sync both directions on link/relink/unlink and via UM create/update
  (verified live) — partner self-identity guards see UI-linked partners.
- EW boundary: 5 dormant Streamlit-only endpoints deleted (stakeholder PII
  leak + unauthenticated SMS dispatch surface); live consumers kept.
- Incident rollback: walks past unstaffed/auto-skipped tiers to the first
  actionable reviewer, or rejects with a clear message; history stamped
  (verified on #91 reject + #34 landing at staffed DED).
- V136 revokes incidents.publish from Dist DC (EOCC doctrine per Authz);
  the three portal-push endpoints now national-guarded (dc→403).
- Warehouses "Stocks" column was hardcoded 0 (audit's one FAKE) → real
  sum(inventory_items.quantity), matches warehouse-ops (verified API==DB).

Note: UserManagementController also carries the previously-uncommitted VAPT
password-policy hardening entangled in the same file (proven 25/25 in the
security retest); committed here deliberately rather than left half-staged.

+++ backend/src/main/java/tz/go/pmo/dmis/preparedness/WarehouseResponse.java
@@ -9,7 +9,7 @@ public record WarehouseResponse(List<WarehouseRow> warehouses, Stats stats) {
     }
 
     public record WarehouseRow(Long id, String name, String cityOrRegion, String address, String zone,
-                               Long capacitySqm, String status, int stocks,
+                               Long capacitySqm, String status, long stocks,
                                String contactName, String contactPhone, Double latitude, Double longitude,
                                String region, String district) {
     }
